requirements.txt
groq 
python-dotenv
streamlit
Gitpython
now 